% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/almanac3-package.R
\docType{package}
\name{almanac3-package}
\alias{almanac3}
\alias{almanac3-package}
\title{almanac3: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/almanac3}
  \item Report bugs at \url{https://github.com/DavisVaughan/almanac3/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
