% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-date.R, R/before-date.R,
%   R/between-dates.R, R/on-date.R
\name{after_date}
\alias{after_date}
\alias{before_date}
\alias{between_dates}
\alias{event-date}
\alias{on_date}
\title{Events related to specific dates}
\usage{
after_date(x, inclusive = FALSE)

before_date(x, inclusive = FALSE)

between_dates(x, y)

on_date(x)
}
\arguments{
\item{x}{\code{[Date(1)]}

A date to mark as an event. For \code{on_date()}, this is
also allowed to be a vector.}

\item{inclusive}{\code{[logical(1)]}

Should \code{x} count as an event?}

\item{y}{\code{[Date(1)]}

A date to mark as an event.}
}
\description{
\itemize{
\item \code{on_date()}: Is the date on \code{x}?
\item \code{before_date()}: Is the date before \code{x}?
\item \code{after_date()}: Is the date after \code{x}?
\item \code{between_dates()}: Is the date between \code{x} and \code{y}?
}
}
