% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-day.R, R/before-day.R, R/between-day.R,
%   R/on-day.R
\name{after_mday}
\alias{after_mday}
\alias{before_mday}
\alias{between_mdays}
\alias{event-mday}
\alias{on_mday}
\title{Events related to days of the month}
\usage{
after_mday(x, inclusive = FALSE, start = TRUE)

before_mday(x, inclusive = FALSE, start = TRUE)

between_mdays(x, y, start = TRUE)

on_mday(x, start = TRUE)
}
\arguments{
\item{x}{\code{[integer(1)]}

A day of the month to mark as an event. For \code{on_mday()}, this is
also allowed to be a vector.}

\item{inclusive}{\code{[logical(1)]}

Should \code{x} count as an event?}

\item{start}{\code{[logical(1)]}

Should counting start from the beginning of the month? If \code{FALSE}, the
event is computed as the \code{x}-th day \emph{from the end} of the month.}

\item{y}{\code{[integer(1)]}

A day of the month to mark as an event.}
}
\description{
\itemize{
\item \code{on_mday()}: Is the date on the \code{x}-th day of the month?
\item \code{before_mday()}: Is the date before the \code{x}-th day of the month?
\item \code{after_mday()}: Is the date after the \code{x}-th day of the month?
\item \code{between_mdays()}: Is the date between the \code{x}-th and \code{y}-th days of
the month?
}
}
