% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-week.R, R/before-week.R,
%   R/between-week.R, R/on-week.R
\name{after_mweek}
\alias{after_mweek}
\alias{before_mweek}
\alias{between_mweeks}
\alias{event-mweek}
\alias{on_mweek}
\title{Events related to weeks of the month}
\usage{
after_mweek(x, inclusive = FALSE, start = TRUE)

before_mweek(x, inclusive = FALSE, start = TRUE)

between_mweeks(x, y, start = TRUE)

on_mweek(x, start = TRUE)
}
\arguments{
\item{x}{\code{[integer(1)]}

A week of the month to mark as an event. For \code{on_mweek()}, this is
also allowed to be a vector.}

\item{inclusive}{\code{[logical(1)]}

Should \code{x} count as an event?}

\item{start}{\code{[logical(1)]}

Should counting start from the beginning of the month? If \code{FALSE}, the
event is computed as the \code{x}-th week \emph{from the end} of the month.}

\item{y}{\code{[integer(1)]}

A week of the month to mark as an event.}
}
\description{
\itemize{
\item \code{on_mweek()}: Is the date on the \code{x}-th week of the month?
\item \code{before_mweek()}: Is the date before the \code{x}-th week of the month?
\item \code{after_mweek()}: Is the date after the \code{x}-th week of the month?
\item \code{between_mweeks()}: Is the date between the \code{x}-th and \code{y}-th weeks of
the month?
}
}
