% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-day.R, R/before-day.R, R/between-day.R,
%   R/on-day.R
\name{after_qday}
\alias{after_qday}
\alias{before_qday}
\alias{between_qdays}
\alias{event-qday}
\alias{on_qday}
\title{Events related to days of the quarter}
\usage{
after_qday(x, inclusive = FALSE, start = TRUE)

before_qday(x, inclusive = FALSE, start = TRUE)

between_qdays(x, y, start = TRUE)

on_qday(x, start = TRUE)
}
\arguments{
\item{x}{\code{[integer(1)]}

A day of the quarter to mark as an event. For \code{on_qday()}, this is
also allowed to be a vector.}

\item{inclusive}{\code{[logical(1)]}

Should \code{x} count as an event?}

\item{start}{\code{[logical(1)]}

Should counting start from the beginning of the quarter? If \code{FALSE}, the
event is computed as the \code{x}-th day \emph{from the end} of the quarter.}

\item{y}{\code{[integer(1)]}

A day of the quarter to mark as an event.}
}
\description{
\itemize{
\item \code{on_qday()}: Is the date on the \code{x}-th day of the quarter?
\item \code{before_qday()}: Is the date before the \code{x}-th day of the quarter?
\item \code{after_qday()}: Is the date after the \code{x}-th day of the quarter?
\item \code{between_qdays()}: Is the date between the \code{x}-th and \code{y}-th days of
the quarter?
}
}
