% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-month.R, R/before-month.R,
%   R/between-month.R, R/on-month.R
\name{after_qmonth}
\alias{after_qmonth}
\alias{before_qmonth}
\alias{between_qmonths}
\alias{event-qmonth}
\alias{on_qmonth}
\title{Events related to months of the quarter}
\usage{
after_qmonth(x, inclusive = FALSE)

before_qmonth(x, inclusive = FALSE)

between_qmonths(x, y)

on_qmonth(x)
}
\arguments{
\item{x}{\code{[integer(1)]}

A month of the quarter to mark as an event. For \code{on_qmonth()}, this is
also allowed to be a vector.}

\item{inclusive}{\code{[logical(1)]}

Should \code{x} count as an event?}

\item{y}{\code{[integer(1)]}

A month of the quarter to mark as an event.}
}
\description{
\itemize{
\item \code{on_qmonth()}: Is the date on the \code{x}-th month of the quarter?
\item \code{before_qmonth()}: Is the date before the \code{x}-th month of the quarter?
\item \code{after_qmonth()}: Is the date after the \code{x}-th month of the quarter?
\item \code{between_qmonths()}: Is the date between the \code{x}-th and \code{y}-th months of
the quarter?
}
}
