% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-week.R, R/before-week.R,
%   R/between-week.R, R/on-week.R
\name{after_qweek}
\alias{after_qweek}
\alias{before_qweek}
\alias{between_qweeks}
\alias{event-qweek}
\alias{on_qweek}
\title{Events related to weeks of the quarter}
\usage{
after_qweek(x, inclusive = FALSE, start = TRUE)

before_qweek(x, inclusive = FALSE, start = TRUE)

between_qweeks(x, y, start = TRUE)

on_qweek(x, start = TRUE)
}
\arguments{
\item{x}{\code{[integer(1)]}

A week of the quarter to mark as an event. For \code{on_qweek()}, this is
also allowed to be a vector.}

\item{inclusive}{\code{[logical(1)]}

Should \code{x} count as an event?}

\item{start}{\code{[logical(1)]}

Should counting start from the beginning of the quarter? If \code{FALSE}, the
event is computed as the \code{x}-th week \emph{from the end} of the quarter.}

\item{y}{\code{[integer(1)]}

A week of the quarter to mark as an event.}
}
\description{
\itemize{
\item \code{on_qweek()}: Is the date on the \code{x}-th week of the quarter?
\item \code{before_qweek()}: Is the date before the \code{x}-th week of the quarter?
\item \code{after_qweek()}: Is the date after the \code{x}-th week of the quarter?
\item \code{between_qweeks()}: Is the date between the \code{x}-th and \code{y}-th weeks of
the quarter?
}
}
