% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on-semester.R
\name{event-semester}
\alias{event-semester}
\alias{on_semester}
\title{Events related to semesters}
\usage{
on_semester(x)
}
\arguments{
\item{x}{\code{[integer]}

A semester to mark as an event.}
}
\description{
\itemize{
\item \code{on_semester()}: Is the date on the \code{x}-th semester?
}
}
