% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-day.R, R/before-day.R, R/between-day.R,
%   R/on-day.R
\name{after_wday}
\alias{after_wday}
\alias{before_wday}
\alias{between_wdays}
\alias{event-wday}
\alias{on_wday}
\alias{on_weekends}
\alias{on_weekdays}
\title{Events related to days of the week}
\usage{
after_wday(x, inclusive = FALSE, start = TRUE)

before_wday(x, inclusive = FALSE, start = TRUE)

between_wdays(x, y, start = TRUE)

on_wday(x, start = TRUE)

on_weekends()

on_weekdays()
}
\arguments{
\item{x}{\code{[integer(1) / character(1)]}

A day of the week to mark as an event. For \code{on_wday()}, this is
also allowed to be a vector. This is also allowed to be a full weekday
string like \code{"Tuesday"}, or an abbreviation like \code{"Tues"}.}

\item{inclusive}{\code{[logical(1)]}

Should \code{x} count as an event?}

\item{start}{\code{[logical(1)]}

Should counting start from the beginning of the week? If \code{FALSE}, the
event is computed as the \code{x}-th day \emph{from the end} of the week.}

\item{y}{\code{[integer(1) / character(1)]}

A day of the week to mark as an event. This is also allowed to be a full
weekday string like \code{"Tuesday"}, or an abbreviation like \code{"Tues"}.}
}
\description{
\itemize{
\item \code{on_wday()}: Is the date on the \code{x}-th day of the week?
\item \code{on_weekends()}: Is the date on a weekend?
\item \code{on_weekdays()}: Is the date on a weekday?
\item \code{before_wday()}: Is the date before the \code{x}-th day of the week?
\item \code{after_wday()}: Is the date after the \code{x}-th day of the week?
\item \code{between_wdays()}: Is the date between the \code{x}-th and \code{y}-th days of
the week?
}
}
