% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-day.R, R/before-day.R, R/between-day.R,
%   R/on-day.R
\name{after_day}
\alias{after_day}
\alias{after_yday}
\alias{before_day}
\alias{before_yday}
\alias{between_days}
\alias{between_ydays}
\alias{event-yday}
\alias{on_day}
\alias{on_yday}
\title{Events related to days of the year}
\usage{
after_day(x, inclusive = FALSE, start = TRUE)

after_yday(x, inclusive = FALSE, start = TRUE)

before_day(x, inclusive = FALSE, start = TRUE)

before_yday(x, inclusive = FALSE, start = TRUE)

between_days(x, y, start = TRUE)

between_ydays(x, y, start = TRUE)

on_day(x, start = TRUE)

on_yday(x, start = TRUE)
}
\arguments{
\item{x}{\code{[integer(1)]}

A day of the year to mark as an event. For \code{on_day()}, this is also
allowed to be a vector.}

\item{inclusive}{\code{[logical(1)]}

Should \code{x} count as an event?}

\item{start}{\code{[logical(1)]}

Should counting start from the beginning of the year? If \code{FALSE}, the
event is computed as the \code{x}-th day \emph{from the end} of the year.}

\item{y}{\code{[integer(1)]}

A day of the year to mark as an event.}
}
\description{
\itemize{
\item \code{on_day()} / \code{on_yday()}: Is the date on the \code{x}-th day of the year?
\item \code{before_day()} / \code{before_yday()}: Is the date before the \code{x}-th day of
the year?
\item \code{after_day()} / \code{after_yday()}: Is the date after the \code{x}-th day of
the year?
\item \code{between_days()} / \code{between_ydays()}: Is the date between the \code{x}-th and
\code{y}-th days of the year?
}
}
\details{
\code{on_day()} is an alias of the less common, but more verbose, \code{on_yday()}.
}
