% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-year.R, R/before-year.R,
%   R/between-year.R, R/on-year.R
\name{after_year}
\alias{after_year}
\alias{after_isoyear}
\alias{after_epiyear}
\alias{before_year}
\alias{before_isoyear}
\alias{before_epiyear}
\alias{between_years}
\alias{between_isoyears}
\alias{between_epiyears}
\alias{event-year}
\alias{on_year}
\alias{on_isoyear}
\alias{on_epiyear}
\title{Events related to years}
\usage{
after_year(x, inclusive = FALSE)

after_isoyear(x, inclusive = FALSE)

after_epiyear(x, inclusive = FALSE)

before_year(x, inclusive = FALSE)

before_isoyear(x, inclusive = FALSE)

before_epiyear(x, inclusive = FALSE)

between_years(x, y)

between_isoyears(x, y)

between_epiyears(x, y)

on_year(x)

on_isoyear(x)

on_epiyear(x)
}
\arguments{
\item{x}{\code{[integer(1)]}

A year to mark as an event. For \code{on_year()}, this is also
allowed to be a vector.}

\item{inclusive}{\code{[logical(1)]}

Should \code{x} count as an event?}

\item{y}{\code{[integer(1)]}

A year to mark as an event.}
}
\description{
\itemize{
\item \code{on_year()}: Is the date on year \code{x}?
\item \code{before_year()}: Is the date before year \code{x}?
\item \code{after_year()}: Is the date after year \code{x}?
\item \code{between_years()}: Is the date between years \code{x} and \code{y}?
}

Additionally, there are equivalent functions for ISO years and
epidemilogical years.
}
