% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-week.R, R/before-week.R,
%   R/between-week.R, R/on-week.R
\name{after_week}
\alias{after_week}
\alias{after_yweek}
\alias{before_week}
\alias{before_yweek}
\alias{between_weeks}
\alias{between_yweeks}
\alias{event-yweek}
\alias{on_week}
\alias{on_yweek}
\title{Events related to weeks of the year}
\usage{
after_week(x, inclusive = FALSE, start = TRUE)

after_yweek(x, inclusive = FALSE, start = TRUE)

before_week(x, inclusive = FALSE, start = TRUE)

before_yweek(x, inclusive = FALSE, start = TRUE)

between_weeks(x, y, start = TRUE)

between_yweeks(x, y, start = TRUE)

on_week(x, start = TRUE)

on_yweek(x, start = TRUE)
}
\arguments{
\item{x}{\code{[integer(1)]}

A week of the year to mark as an event. For \code{on_week()}, this is also
allowed to be a vector.}

\item{inclusive}{\code{[logical(1)]}

Should \code{x} count as an event?}

\item{start}{\code{[logical(1)]}

Should counting start from the beginning of the year? If \code{FALSE}, the
event is computed as the \code{x}-th week \emph{from the end} of the year.}

\item{y}{\code{[integer(1)]}

A week of the year to mark as an event.}
}
\description{
\itemize{
\item \code{on_week()} / \code{on_yweek()}: Is the date on the \code{x}-th week of the year?
\item \code{before_week()} / \code{before_yweek()}: Is the date before the \code{x}-th week of
the year?
\item \code{after_week()} / \code{after_yweek()}: Is the date after the \code{x}-th week of
the year?
\item \code{between_weeks()} / \code{between_yweeks()}: Is the date between the \code{x}-th and
\code{y}-th weeks of the year?
}
}
\details{
\code{on_week()} is an alias of the less common, but more verbose, \code{on_yweek()}.
}
