% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-in.R
\name{event_in}
\alias{event_in}
\title{Is \code{x} an event?}
\usage{
event_in(x, event)
}
\arguments{
\item{x}{\code{[Date]}

A vector of dates.}

\item{event}{\code{[event]}

An event.}
}
\description{
\code{event_in()} checks if \code{x} is in the set of dates defined by \code{event}.
}
\examples{
event_in("2019-01-01", on_mday(1))

event_in("2019-01-01", on_mday(1) & on_wday("Mon"))

}
