% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch-add.R
\name{sch_add}
\alias{sch_add}
\title{Add an event to a schedule}
\usage{
sch_add(x, event, name = NULL)
}
\arguments{
\item{x}{\code{[schedule]}

A schedule.}

\item{event}{\code{[event]}

An event.}

\item{name}{\code{[character(1) / NULL]}

A name for the event. If left as \code{NULL}, a default name will be created.}
}
\description{
\code{sch_add()} allows you to add an \code{event} to a schedule, optionally providing
a \code{name} for the event.
}
\examples{
# Labor Day = First Monday in September
on_labor_day <- on_month("Sep") & on_wday("Mon") & on_mweek(1)

# Christmas = December 25th
on_christmas <- on_month("Dec") & on_mday(25)

sch <- schedule()
sch <- sch_add(sch, on_labor_day, "Labor Day")
sch <- sch_add(sch, on_christmas, "Christmas")

sch

sch_in("2019-01-01", sch)

sch_in("2019-12-25", sch)

}
