% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch-retrieve.R
\name{sch_retrieve}
\alias{sch_retrieve}
\title{What events does \code{x} fall on?}
\usage{
sch_retrieve(x, schedule)
}
\arguments{
\item{x}{\code{[Date]}

A vector of dates.}

\item{schedule}{\code{[schedule]}

A schedule.}
}
\description{
\code{sch_retrieve()} returns the events that \code{x} falls on in the \code{schedule}.
}
\examples{
on_25th <- on_mday(25)
on_dec <- on_month("Dec")

sch <- schedule()
sch <- sch_add(sch, on_25th, "twenty_fifth")
sch <- sch_add(sch, on_dec, "december")

sch_retrieve(c("2019-01-01", "2019-01-25", "2019-12-25"), sch)

}
