% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule}
\alias{schedule}
\title{Create a new schedule}
\usage{
schedule()
}
\description{
\code{schedule()} creates a new empty schedule. Add events to the schedule with
\code{\link[=sch_add]{sch_add()}}.
}
